build: add `python-semantic-release` as a dev dependency
This dependency introduced additional dependencies that failed the `phylum-ci` project thresholds for license risk: `docutils@0.18.1` and `python-gitlab@3.3.1`. `python-gitlab` uses the LGPLv3, which was classified as critical. This appears to be a mis-classification as the license allows for the uses expected in this project. `docutils` uses a public domain license, but also references several other licenses from code it makes use of. One of them is the GPL-3 license, which was classified as high. This also appears to be a mis-classification as the license allows for the uses expected in the project. Both packages were suppressed for their license risk findings.