fix: use `phylum-bot` account instead of a personal account
The previous GitHub Personal Access Token (PAT) used in the `Release` workflow was created with a short lifetime and from a personal account (@maxrake). This was not sustainable and was replaced by a company controlled bot/service account (@phylum-bot). This account was configured for code signing and vigilant mode. The `Production` deployment environment and `Release` workflow were updated to include the secrets necessary to make use of the code signing to ensure verified commits and tags.

closes #33