!feat: default to analyzing new dependencies only (#53)
Most users of these integrations will already have an existing code base and will not want to "fail their CI builds" for existing dependencies while they work to clean them up. This change reverses the logic of the `--new-deps-only` option so that users specify an option when they want to analyze all dependencies. The default will now be to analyze only newly added ones.

BREAKING CHANGE: The option to analyze `--new-deps-only` was removed and replaced with one that has the opposite meaning: `--all-deps`