perf: optimize Docker image
The image base layer was switched from one based on Debian to one based on Alpine. The compressed image size was reduced from 84MB to 35MB. The expanded image size was reduced from 245MB to 96MB. The image acquisition process in GitLab was reduced from ~16s to ~9s and the total run time was reduced from ~26s to ~19s. The Dockerfile now makes use of BuildKit for caching purposes. It can be used to create images used for layer caching. It is also making use of a cache type build mount in order to re-use already built wheels...which cuts the build time for cross compiled builds by ~75%.

Closes #48