fix: docker image tags are formatted differently
The Docker workflow makes use of steps that expect the `phylum-ci`
release version to include the leading `v`, like checking out the repo
and getting the Python wheel from the latest release. This is due to how
the GitHub releases are referenced by the tag name, which includes the
leading `v`. However, the Docker image tags created by both the Release
and Docker workflows expect the version to be provided without the `v`,
based on standard tag formats in use today.

This change ensures the tags are created in the same format, regardless
of how or where the images are built. The format is:

`<phylum-ci version>-CLIv<Phylum CLI version>`

where both version component fields are expected to be semantic version
identifiers *without* a leading `v`.

Some additional formatting and documentation updates were made.

Closes #66