ci: update Phylum Analyze PR action to v2
The individual thresholds were removed since they are set at the project level. Additionally, the `license` threshold was reduced to 20 in the project settings since group accounts do not allow for issue suppression at the moment and there are a handful of false positive issues related to the license risk domain, the lowest of which is reporting as a 20.