fix: detect lockfile changes in GitHub PRs (#73)
Adding the $GITHUB_WORKSPACE as a safe directory in `git` is necessary to ensure the GitHub Container action will function properly with the git repository checked out on the runner. However, the recommended workaround was applied too late in the class initialization. The parent class sets the instance predicate attribute for `is_lockfile_changed` in the `__init__` method. Therefore, the workaround needs to be applied in the child class *prior* to this.

Closes #72