docs: create exclusive directory for Integrations docs to sync properly (#80)
Providing a directory with a mix of markdown files, some to be synced and others that shouldn't be, was causing the Readme.com sync task to fail the `Release` workflow. This change provides an exclusive directory to house only those markdown files that are destined for the Integrations section of the Phylum docs page. It also includes a Readme.txt file (not .md!) to explain the situation.

Closes #79