ci: automate dependency updates and version alerts
Dependabot is enabled for version alerts and update PR for GitHub Actions, Dockerfile, and major Python package versions. A separate "Poetry Update" workflow is added to handle the rest of the Python package versions. This workflow was modeled after the one in the `phylum-dev/cli` repo (which is for Cargo updates).

Closes #17