build: remove the caps from project dependencies
This change is an attempt to follow the advice of folks who have thought about dependency caps in Python at a deep level. See https://iscinumpy.dev/post/bound-version-constraints/ and https://bernat.tech/posts/version-numbers/ for starters. The basic advice is that upper bound version caps are to be avoided. The use of lockfiles is still the best practice for applications. Poetry continues to present a problem with removing the cap for the Python version itself. Indeed, attempting to remove it for this project resulted in a regression of package versions specified in the lockfile. So much so that the overall Phylum project score went from 83 to ~23 (some packages were still processing).