feat: render markdown output in terminal for pre-commit and no-CI environments
So far, the CI environments that have been implemented allow for output, in the form of review comments, to be posted as rendered markdown. The environments that don't use CI - pre-commit and 'no-CI' so far - display their output in the terminal. Instead of writing separate output for these environments, a conversion utility library is used to render the existing markdown output in the terminal. Additionally, the labels for these environments were shortened to be more readable...in both the output as a link and the Phylum UI in the label dropdown menu.