fix: `Issue Summary` data missing for vulnerability domain (#99)
Naming things is hard. Keeping names for risk domains consistent across the various places they are referenced is even harder, it appears. There are still some differences in the API response data for the vulnerability/vulnerabilities risk domain. There is also still a difference for `malicious` and `malicious_code`. These differences are accounted for in the `RiskDomain` dataclass, but they were not accounted for the different locations these names show in the json response from the CLI: project level threshold name, individual package `riskVectors` name, and domain name for an issue in an individual package. This change attempts to account for each of the potential names in each of the places they could be referenced.

Closes #98