build: add `rich-codex` as a dev dependency
This is really only a dependency needed for release. It could also have been left out of the `poetry` dependencies and installed directly during the release workflow instead, but this method allows for keeping tabs on the project via Phylum analysis of the lockfile. It also ensures that the pre-commit command run as part of the semantic-release will be available, regardless of how/where the semantic-release is initiated.