feat!: provide a Docker image with glibc instead of musl libc (#104)
The basis for this change is due to how the Linux platform support in the CLI was changed from `unknown-linux-musl` to `unknown-linux-gnu` starting with v3.8.0-rc2, changing the artifact names available to download and install in a non-backwards compatible manner. This is also a breaking change in this repository as the `Dockerfile` had to migrate away from one based on a `python:3.10-alpine` base image to one based on `python:3.10-slim-bullseye`. The resulting docker image is ~113MB compressed (up from ~38MB) and ~327MB uncompressed (up from ~106MB). Documentation was also updated to account for the new minimum supported version of the CLI.

Closes #103

BREAKING CHANGE: Versions of the CLI older than v3.8.0-rc2 are no longer possible to install on Linux systems with the `phylum-init` script.