fix: script options auto update can't find package
The last release showed that the script options documentation auto-
update feature was broken. Instead of creating an updated SVG file(s)
for the script options, a stack trace was produced as output instead,
with the following error:

`ModuleNotFoundError: No module named 'phylum'`

The belief is that the `python-semantic-release` package, which uses the
`invoke` package under the hood to run commands, issued a command
outside of the `poetry` environment and therefore did not have access to
the installed local `phylum` package.

This change attempts to correct for that oversight by ensuring the
command is run within the poetry environment where the `phylum` package
is installed. It is an _attempt_ because it still may not work and it
won't be possible to tell for sure until the next release. If that
happens, there are alternative solutions available to pursue.