fix: script options auto update still can't find package (#108)
The last release showed that the script options documentation auto-
update feature is still broken. Instead of creating an updated SVG
file(s) for the script options, a stack trace was produced as output
instead, with the following error:

`ModuleNotFoundError: No module named 'phylum'`

A fix was attempted with #107 but could not be fully verified until the
release was made. The belief now is that the `python-semantic-release`
package does not provide enough control over running commands with the
`pre_commit_command` option in the config file.

This change attempts to run the command in the release workflow, as it's
own step, *before* the Python semantic-release step. Everything needed
to run the `rich-codex` command exists in the workflow environment at
the point the step is run. The code for the `python-semantic-release`
package was examined and it appears to properly handle
additional/changed files with the `include_additional_files` config
option. It would have been nice to have everything specified in the
config, but this approach should work.