ci: allow for TestPyPI publish failures in release
The release workflow step that publishes the package to the TestPyPI
registry should be allowed to continue on error since there are times
when it is expected to fail (e.g., when re-running the workflow). The
following step will still provide the gating function for the package
verification on the TestPyPI registry. That is, if the verification step
fails, the job will still fail.