ci: install current project in release workflow
The release workflow needs the current project (`phylum`) to be
installed now that it gets used by the `rich-codex` tool to update the
script options documentation automatically. Making this change may mean
slightly more time is taken to install the project in CI. It may also
affect the Python cache used in the workflow.