feat: don't require serial processing of pre-commit hook (#115)
It is more desirable to not `require_serial` configurations in hooks, so that they may be run in a parallel fashion. Allowing for this `pre-commit` configuration also means that the pre-requisite check for the pre-commit environment should not exit with an error when the lockfile was not included in the extra arguments. Instead, it should simply exit...with message(s) to make the situation leading up to that exit more clear.

Closes #114