ci: delete `issues.yml` (#117)
* ci: delete `issues.yml`

* ci: replace `ADD_TO_PROJECT_PAT` with `GH_RELEASE_PAT`

The `ADD_TO_PROJECT_PAT` is exposed as an organization secret and it will be removed as such for this repository soon. The `GH_RELEASE_PAT` will do the same job, with less permissions. The `GH_RELEASE_PAT` is currently an environment secret (for the `Production` environment) but will be moved to a repository secret so that the auto_updates workflow will continue to function without needing a review/approval from the rules in place on the `Production` environment.

Co-authored-by: Charles Coggins <maxrake@users.noreply.github.com>