refactor!: remove legacy CLI path logic
The minimum supported CLI version is currently v3.8.0-rc2 and enough
time has passed that CLI versions prior to v2.2.0 are no longer expected
to exist in the wild. That is when the Phylum config and binary paths
changed, to adhere to the XDG Base Directory Spec.

This change removes support for use of these legacy CLI version paths.

BREAKING CHANGE: CLI installs prior to v2.2.0 are no longer supported.