fix: account for shallow fetch in Azure Pipelines integration (#135)
The documentation has been updated to be more clear about how to disable the shallow fetch option. There is also a better error message reminding users that shallow fetch should be disabled, with a link to more information about how to do so.

There is also an effort to provide more detail around Azure token failures. The places where having an Azure Token and having the correct permissions on that token have been updated to provide a more detailed error message when those criteria are not met.

Closes #134