docs: update GitLab CI integration docs
The documentation has been updated to match the changes in the integration. It accounts for the possibility of running the integration in the context of both merge request pipelines and now branch pipelines. The Job Control section was updated to better show how to enable one or the other pipeline, or both...but without creating duplicate running pipelines.