feat: support GitLab branch pipelines (#137)
The ability to recognize when operating in the context of a branch pipeline was encapsulated with a class property. This property was only added to the GitLab integration at this time. It may make sense to extract this property up to the base class, if and when additional environments are modified to work in the same way. The logic was updated to account for the differences between branch pipelines and merge request pipelines, with the preference (order of operations) tending towards merge request pipelines. Output from the analysis will be posted to the log file of either pipeline type, but posting a note to the MR will only happen now when operating in a merge request pipeline. Due to this change, it is now no longer required to have a GitLab token available at all times since it is not required for branch pipelines.

The documentation has been updated to match the changes in the integration. It accounts for the possibility of running the integration in the context of both merge request pipelines and now branch pipelines. The Job Control section was updated to better show how to enable one or the other pipeline, or both...but without creating duplicate running pipelines.

The scant documentation that GitLab CI provides for the `CI_COMMIT_BEFORE_SHA` variable was not exactly correct. At least, it didn't tell the whole story. After direct, manual trial and error, its true nature was discovered. That has also been documented and accounted for in comments and code.

Closes #132