feat: allow for creating projects
The change adds options to specify a project and, optionally, a group
to go with that project. The group option is ignored if the project
option is not populated.

When the project option is provided, a project will be created,
overwriting any existing `.phylum_project` file. If a group is provided,
it will also be created. If the provided project and group already
exist, they will be used.

Output is provided to inform users about the new options when the
integration fails as a result of a missing `.phylum_project` file.

Closes #136