fix: account for no `.phylum_project` file
The `.phylum_project` file may not exist in the case where it didn't
exist initially and then it was not created because the project (with
optional group) provided as input options already existed. Therefore, it
is no longer possible to assume that the `.phylum_project` will always
be present.

This comes into play for the following situations:
* when performing an `analyze` command
* when finding the group name to create the project url

The `phylum_group` class property has been updated to account for the
various sources a group can be defined. The `phylum analyze` command was
updated to ensure the provided project and optional group name is used.