feat: allow for creating projects (#139)
This change adds options to specify a project and, optionally, a group
to go with that project. The group option is ignored if the project
option is not populated.

When the project option is provided, a project will be created,
overwriting any existing `.phylum_project` file. If the provided project
and group already exist, they will be used.

Effective project name and group name class properties were added with
logic so that the values for project name and group name can either come
from the command line options or the `.phylum_project` file. Mixing of the
two is not allowed. Preference is given to command line options.

The `.phylum_project` file may not exist in the case where it didn't
exist initially and then it was not created because the project (with
optional group) provided as input options already existed. Therefore, it
is no longer possible to assume that the `.phylum_project` will always
be present.

This comes into play for the following situations:
* when performing an `analyze` command
* when finding the group name to create the project url

These situations were addressed by using the new class properties.

Output is provided to inform users about the new options when the
integration fails as a result of a missing `.phylum_project` file.

Closes #136

Co-authored-by: Kyle Willmon <kyle@phylum.io>