fix: sanitize user input to guard against possible cmd injection
This change makes use of the `shlex` module to `quote` the parts of any
command line that can possibly come from user supplied input. The
command line is then `split` with the same module to ensure proper and
sanitized tokenization when supplied to a `subprocess.run` call.

The `shlex` module is only designed for Unix shells. The `shlex.quote()`
function is not guaranteed to be correct on non-POSIX compliant shells
or shells from other operating systems such as Windows. Therefore, the
documentation and PyPI package classifiers were updated to make that
operating limitation more obvious.

Fixes #143