fix: sanitize user input to guard against possible cmd injection
This change backs out the use of the `shlex` module to quote the parts
of any command line that can possibly come from user supplied input.
Explicit lists of command line arguments are used instead. Every element
of each list is a single string, regardless of whitespace contained
within. This is true for both static string literals and variables,
whether they came from user input or not. Therefore, each element in the
list will be considered a separate token/argument when supplied to a
`subprocess.run` call. Every instance of `subprocess.run` was reviewed
and updated to this new format.

Fixes #143