fix: sanitize user input to guard against possible cmd injection (#144)
This change ensures that explicit lists of command line arguments are
used for every input to `subprocess.run` instead of splitting strings. Every
element of each list is a single string, regardless of whitespace contained
within. This is true for both static string literals and variables,
whether they came from user input or not. Therefore, each element in the
list will be considered a separate token/argument when supplied to a
`subprocess.run` call. Every instance of `subprocess.run` was reviewed
and updated to this new format.

Fixes #143