docs: Provide more hints about using the SVG files
During conversations with a potential customer, feedback was provided to say that the top-level README was not clear enough on the use of SVG files to show the current script options help output. That is, they did not realize that the "images" were actually SVG files that, when clicked on and opened in raw form in the browser, would allow for search and copy/paste functionality. This change makes that hint more obvious by using a markdown quote box, with a bold **HINT** preface. Additionally, the hint is added to all the places the SVG files are displayed.