build: use poetry plugin for export command
The `export` command is now provided as a plugin. It is installed by default for the `poetry` v1.2 release, but future releases will deprecate this automatic install. Installing the plugin directly now does not harm the function of the plugin or cause any errors given that it may already be installed. This change is getting ahead of the future deprecation.