ci: use single test rollup job
This change aggregates all the test jobs into the "test rollup" job so that there will only be one job name that needs to be specified in the repository settings for branch protection and required status checks. The advantage of doing this is that the name of the status check will not change when the version of Python the underlying jobs use changes.