docs: update the Azure Pipelines documentation to support GitHub repos
The docs have been modified to account for both Azure Repos Git and
GitHub as build/triggering repositories. A more detailed error message
about the use of personal access tokens (PATs) was added for the GitHub
integration and repurposed for the Azure Pipelines integration. Some
formatting, refactoring, and typo corrections was also applied.