refactor: use the long form option for label
Starting with CLI v3.12.0, `--label` was added as a long form option to the `analyze` command. This new form is preferred since it is more descriptive. The minimum supported version constants were updated to reflect this. The version is just slightly newer than the one where openssl signatures were introduced (v3.10.0) and so this change is considered relevant and timely to coincide with the changes made to support openssl RSA signature verification for CLI artifact downloads.

BREAKING CHANGE: CLI installs prior to v3.12.0 are no longer supported.