chore: specify GitHub REST API version (#166)
As of 28 November 2022, the GitHub REST API is versioned with a new
calver scheme. The recommendation from GitHub is to update all
integrations that make use of this API to be explicit about the version
in use. This change does just that for the `phylum-ci` repo. Since all
the calls work with the same version at this point, a constant was added
to allow for a single source for changing to new/different versions in
the future.

References:
* [Blog Post](https://github.blog/2022-11-28-to-infinity-and-beyond-enabling-the-future-of-githubs-rest-api-with-api-versioning/)
* ["API Versions" documentation](https://docs.github.com/rest/overview/api-versions)