feat!: compare GitLab branch pipeline changes to default branch
This change makes the behavior of the GitLab integration match that of
the Azure Pipelines integration.

BREAKING CHANGE: For GitLab branch pipelines, the analyzed dependencies
are now determined by comparing the lockfile in the branch to the
default branch instead of the previous commit that ran in that branch
pipeline. All dependencies will be analyzed when the branch pipeline is
run on the default branch.