feat!: support Azure Pipelines CI triggers (#173)
The Azure Pipelines integration was updated to recognize and support working in CI triggered (pushes to a branch) pipeline contexts. Azure Pipelines does not provide a means to determine the previous commit that triggered a build in the same branch pipeline like GitLab does. Instead, a design review amongst stakeholders was held to decide which of several options to pursue:

1. Treat the commits as singular...applying an `--all-deps` strategy always for CI triggered pipelines
2. Get the merge base between the commit and the default branch and use that to find diffs
    * When already on the default branch, provide info/issues on all the deps
    * Use `refs/remotes/origin/HEAD` to determine the default branch
3. Use the ADO API for finding the information needed to determine the previous build for the branch
    * PRO: might provide the info needed w/o requiring a deep clone/checkout
    * CON: requires an ADO Token and adds network call overhead

Option 2 was selected. It turns out that Azure Pipelines repository checkout steps do not set the remote HEAD ref. An alternate approach was implemented with a git command to automatically set it.

The GitLab integration was updated to make it's behavior match that of the Azure Pipelines integration. It now compares branch pipeline changes to the default branch.

Additional actions taken:

* Extract common git functionality into a new module
  * The git functionality that is repeated and not dependent on context or expected to cause exceptions was pulled out into a new `git` module
  * Add new function to provide the default branch for the current repository
  * Add new function to set the remote HEAD ref
* Expose `--all-deps` and `--force-analysis` options as class properties
  * Creating new class properties for the state of the `--all-deps` and `--force-analysis` options will allow for changing those values/state for situations where it makes sense to analyze all dependencies (but the user didn't explicitly request it).
* Simplify the `str.removeprefix()` replacement logic
* Add `backports-cached-property` project dependency 
  * This is a drop in backport for the `@functools.cached_property` decorator available in the standard library starting with Python 3.8
  * It will use the standard library version when available
* Use `cached_property` decorator to avoid needless computations
* Update the Phylum project file to the latest version
* Refresh the dependencies in the lockfile
* Update documentation to match the changes made
* Refactor and format throughout

Closes #149

BREAKING CHANGE: For GitLab branch pipelines, the analyzed dependencies are now determined by comparing the lockfile in the branch to the default branch instead of the previous commit that ran in that branch pipeline. All dependencies will be analyzed when the branch pipeline is run on the default branch.