fix: issue summary entries repeated in output
It turns out that the possible risk domain names are mostly the same between how they are referenced between the project and package levels. The list of potential names was getting added to the list of flags, *without* checking for duplicates first. This was corrected by using a set instead of a list for the potential names. Additionally, the difference between `vulnerabilities` and `vulnerability` was confirmed to have been corrected in phylum-dev/api#424 and thus that entry was updated here as well. So, the only remaining difference is between `malicious_code` and `malicious`, which is being tracked with phylum-dev/api#499

Fixes #174