chore: include GITHUB_TOKEN in release workflows for docker builds
In adding the feature to increase the rate limits for GitHub API calls,
a few of the `phylum-ci` repository workflows were updated to take
advantage of the fact that the `GITHUB_TOKEN` is already available in
those workflows. However, two updates were missed: the release and
docker workflows. This change includes them.