fix: link to Phylum UI project clipped in logs
This change addresses a formatting anomaly in the terminal/log output
where the link to the way project in the Phylum UI was getting clipped
and then no longer functional as a clickable link in the environments
where links are rendered as clickable.

The `connect-markdown-renderer` and `rich` libraries are responsible for
rendering the markdown in a terminal. The width is automatically
determined (maybe via environment variable) and for some CI environments
the width is smaller than the link, causing it to get clipped.

The library method in use now does not allow for specifying different
behavior. If it did, the easiest solution would be to pass
`soft_wrap=True` through to the instantiation of the `Console` class in
`rich`. Instead, the link display was removed from the markdown for the
output that gets posted to logs and consoles. It remains in the markdown
for the output that gets posted to CI environments where markdown is
supported (currently all of them).