feat: add support for Bitbucket pipelines
This change adds an integration for the Bitbucket pipelines CI
environment. Specifically, it adds support for the Bitbucket Cloud
ecosystem, which is their CI/CD as a service offering for repositories
hosted at <https://bitbucket.org/>.

This integration detects and handles both branch pipelines and pull
request (PR) pipelines. When in the PR pipeline context, the results
will be written to logs and also as a comment on the PR. PR comments
in Bitbucket do not allow for HTML tags from markdown input and will
display the tags as text instead of a rendered blend of HTML and
Markdown.

It was discovered that not only was Bitbucket affected by the inline
HTML tags, but also the integrations and code paths that do not rely on
posting a PR comment. That included: `pre-commit`, the local use case
`CI-None`, `Azure` and `GitLab` for their branch pipeline output, and
_all_ the log output for integrations with pull request pipeline output
(GitHub, GitLab, Azure, Bitbucket).

It was decided to remove the HTML tags entirely from all Phylum
comments. That means the text will be displayed normally instead of in a
collapsible element. The text was also modified to be more succinct.
This means that for some integrations there will be more text shown by
default (vs. in a collapsible element). It also means that all
integrations will have the same output "look and feel" for failures.

Closes #189