fix: duplicate PR comments are possible (#199)
This change ensures that comments are constructed in a deterministic and
consistent manner. Any packages with risk findings will be sorted, first
by the package name and version. Then, the individual issues for a given
package are sorted by severity level (critical --> info), risk domain,
and title. The result of these changes should ensure that PR comments
are not posted again when the only difference from the previous
`phylum-ci` generated comment is the order of the content within the
comment.

Fixes #193