fix: collisions possible when getting previous lockfile object
The `git rev-parse` command was incorrectly using the `.name` part of the given lockfile's `Path` object, which simply returns the basename. This was a bug because it is possible for lockfiles with different paths (e.g., in different subdirectories) to have the same basename (e.g., `requirements.txt`, `lvl2/requirements.txt`, and `lvl2/lvl3/requirements.txt`). Instead, the relative path to the lockfile, as taken from the current working directory, should be used. This relative path will work for the `git rev-parse` command even when the CWD is a subdirectory since `git` will traverse up to the top-level directory (e.g., the one containing a `.git` directory) and use that as the reference point. The `__repr__` form of a `Lockfile` object provides exactly this relative path desired. The `repr` form was updated to be used in all the places it makes sense.