fix: `is_lockfile_changed` property not always set
The `is_lockfile_changed` property should be set (not `None`) before the a lockfile's `new_deps` property is used. This condition was violated for the times when the `--force-analysis` flag was provided but the `--all-deps` flag was not. The result is that no newly added dependencies were being recognized. This change ensures each lockfile's change status is updated for that flag combination. It does so in a way that continues to allow the integration to bail early when there are no changes to any lockfile and generally run faster by avoiding unnecessary operations.