refactor: use source branch consistently in PR/MR labels
While testing the changes in this branch across the various CI environments, it was noticed that the GitLab integration crafted a label for analysis in a different way than the integrations for the other environments. It was using the Merge Request _title_ where the others are using the source branch name. The GitLab label for MRs was updated to also use the source branch.