feat: use correct PR number in ADO labels for GitHub repos
The Azure Pipelines integration supports repositories hosted in either Azure Repos Git or GitHub. The label created for PRs was using the PR ID for both sources. However, this number is meaningless for repos hosted in GitHub. In those cases, there is a different environment variable provided that contains the equivelant PR number.