refactor: use `shlex.join()` function
The `shlex.join()` function was introduced in Python 3.8 and returns a
shell-escaped string from a split command and is used to protect against
injection vulnerabilities. It is the inverse of `shlex.split()` and
simplifies the previous method of manually using `shlex.quote()` for
every piece of the command line input.