ci: add `refurb` pre-commit hook (#221)
This change adds `refurb` to the QA checks and updates the code to
adhere to it's findings. From the [refurb
README](https://github.com/dosisod/refurb#why-does-this-exist), the tool
is "heavily inspired by clippy, the built-in linter for Rust."

The defaults are used, except for specifying the minimum supported
Python version (currently 3.8) so that changes won't be suggested that
only exist in newer versions of Python. Additionally, the format when
using the pre-commit hook is set to `github`. Even though this is
slightly harder to parse when running the hook locally, the benefit
outweighs that cost in the way annotations will be added to PRs.